+++ arctic/store/versioned_item.py
@@ -1,7 +1,7 @@
 from collections import namedtuple
 
 
-class VersionedItem(namedtuple('VersionedItem', ['symbol', 'library', 'data', 'version', 'metadata'])):
+class VersionedItem(namedtuple('VersionedItem', ['symbol', 'library', 'data', 'version', 'metadata', 'host'])):
     """
     Class representing a Versioned object in VersionStore.
     """
@@ -12,8 +12,8 @@ class VersionedItem(namedtuple('VersionedItem', ['symbol', 'library', 'data', 'v
         return str(self)
 
     def __str__(self):
-        return "VersionedItem(symbol=%s,library=%s,data=%s,version=%s,metadata=%s" % \
-            (self.symbol, self.library, type(self.data), self.version, self.metadata)
+        return "VersionedItem(symbol=%s,library=%s,data=%s,version=%s,metadata=%s,host=%s)" % \
+            (self.symbol, self.library, type(self.data), self.version, self.metadata, self.host)
 
 
 ChangedItem = namedtuple('ChangedItem', ['symbol', 'orig_version', 'new_version', 'changes'])