+++ tests/unit/store/test_version_item.py
@@ -8,10 +8,11 @@ def test_versioned_item_str():
                          library="ONEMINUTE",
                          data=pd.DataFrame(),
                          version=1.0,
+                         host='myhost',
                          metadata={'metadata': 'foo'})
 
     expected = "VersionedItem(symbol=sym,library=ONEMINUTE," + \
-               "data=<class 'pandas.core.frame.DataFrame'>,version=1.0,metadata={'metadata': 'foo'}"
+               "data=<class 'pandas.core.frame.DataFrame'>,version=1.0,metadata={'metadata': 'foo'},host=myhost)"
     assert str(item) == expected
     assert repr(item) == expected
 
@@ -21,7 +22,8 @@ def test_versioned_item_str_handles_none():
                          library=None,
                          data=None,
                          version=None,
-                         metadata=None)
+                         metadata=None,
+                         host=None)
 
     assert str(item)
 
@@ -31,5 +33,6 @@ def test_versioned_item_metadata_dict():
                          library="test_lib",
                          data=None,
                          version=1.2,
-                         metadata=None)
+                         metadata=None,
+                         host=None)
     assert(item.metadata_dict() == {'symbol': 'test', 'library': 'test_lib', 'version': 1.2})